Format results from list objects in APA style

This function detects the type of statistical results stored in a list object and redirects to the appropriate APA formatting method. Currently supports ez::ezANOVA results (apa.ezANOVA).

Usage

# S3 method for class 'list'
apa(obj, ...)

Arguments

obj

A list object containing statistical results

...

Additional arguments passed to the appropriate method

Value

A formatted string with statistical results in APA style